+++ backend/src/main/java/com/mcm/backend/app/api/controllers/users/user/UserMeController.java
@@ -1,0 +1,5 @@
+package com.mcm.backend.app.api.controllers.users.user;
+
+public class UserMeController {
+    // TODO Implement
+}
