+++ backend/src/main/java/com/mcm/backend/app/api/core/HealthCheckController.java
@@ -1,4 +1,4 @@
-package com.mcm.backend.app.routes.core;
+package com.mcm.backend.app.api.core;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
